Pavilion authorization added.

+++ back/app/Http/Controllers/PavilionController.php
@@ -9,11 +9,13 @@ class PavilionController extends Controller
 {
     public function index()
     {
+        $this->authorize('viewAny',Pavilion::class);
         return response()->json(Pavilion::all());
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create',Pavilion::class);
         return response()->json(Pavilion::create([
             "title"=>$request->input('title'),
             "price_per_day"=>$request->input('price_per_day'),
@@ -23,11 +25,13 @@ class PavilionController extends Controller
 
     public function show($id)
     {
+        $this->authorize('view',Pavilion::class);
         return response()->json(Pavilion::find($id));
     }
 
     public function destroy($id)
     {
+        $this->authorize('destroy',Pavilion::class);
         return response()->json(Pavilion::destroy($id));
     }
 }
